+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## [0.2.1] - 2026-07-19
+- **Bug Fixes:** Migration naming issue.
+- **Bug Fixes:** Contact search dropdown slot was not correctly passed to the component.
+
 ## [0.2.0] - 2026-07-19
 - **Contact:** Implement contact relation between customer and contacts from contacts app.
 - **Address book:** The CRM Inits now a address book for crm contacts.